+++ .eslintrc.json
@@ -1,44 +1,48 @@
 {
-    "env": {
-        "es2021": true,
-        "node": true
-    },
-    "extends": [
-        "eslint:recommended",
-        "plugin:@typescript-eslint/recommended"
-    ],
-    "overrides": [
-    ],
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "ecmaVersion": "latest",
-        "sourceType": "module"
-    },
-    "plugins": [
-        "@typescript-eslint",
-        "unused-imports"
-    ],
-    "rules": {
-        "comma-dangle": ["warn", "always-multiline"],
-        "comma-spacing": ["warn", { "after": true }],
-        "eqeqeq": ["error", "always"],
-        "indent": ["warn", 2, { "SwitchCase": 2 }],
-        "key-spacing": ["warn", { "afterColon": true }],
-        "max-len": ["warn", { "code": 120 }],
-        "no-extra-boolean-cast": "off",
-        "no-nested-ternary": "error",
-        "no-multi-spaces": "warn",
-        "no-redeclare": "off",
-        "@typescript-eslint/no-redeclare": "error",
-        "no-unused-vars": "off",
-        "@typescript-eslint/no-empty-interface": ["off"], 
-        "@typescript-eslint/no-unused-vars": ["warn", { "args": "after-used", "ignoreRestSiblings": true }],
-        "object-curly-spacing": ["warn", "always"],
-        "quotes": ["warn", "single"],
-        "semi": ["error", "always"],
-        "space-in-parens": ["warn", "never"],
-        "spaced-comment": ["warn", "always", { "markers": ["/"] }],
-        "unused-imports/no-unused-imports": "warn",
-        "eol-last": 2
+  "root": true,
+  "extends": [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/recommended"
+  ],
+  "plugins": [
+    "unused-imports"
+  ],
+  "env": {
+    "es2021": true,
+    "node": true
+  },
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  },
+  "rules": {
+    "comma-dangle": ["warn", "always-multiline"],
+    "comma-spacing": ["warn", { "after": true }],
+    "eqeqeq": ["error", "always"],
+    "indent": ["warn", 2, { "SwitchCase": 2 }],
+    "key-spacing": ["warn", { "afterColon": true }],
+    "max-len": ["warn", { "code": 120 }],
+    "no-extra-boolean-cast": "off",
+    "no-nested-ternary": "error",
+    "no-multi-spaces": "warn",
+    "no-redeclare": "off",
+    "@typescript-eslint/no-redeclare": ["error"],
+    "no-unused-vars": "off",
+    "@typescript-eslint/no-empty-interface": ["off"], 
+    "@typescript-eslint/no-unused-vars": ["warn", { "args": "after-used", "ignoreRestSiblings": true }],
+    "object-curly-spacing": ["warn", "always"],
+    "quotes": ["warn", "single"],
+    "semi": ["error", "always"],
+    "space-in-parens": ["warn", "never"],
+    "spaced-comment": ["warn", "always", { "markers": ["/"] }],
+    "unused-imports/no-unused-imports": "warn",
+    "eol-last": 2
+  },
+  "overrides": [
+    {
+      "files": ["*.ts", "*.tsx"],
+      "parser": "@typescript-eslint/parser"
     }
+  ]
 }
