+++ .vscode/settings.json
@@ -1,11 +1,13 @@
 {
     "editor.codeActionsOnSave": {
-        "source.fixAll.eslint": true,
+      "source.fixAll.eslint": "explicit"
     },
-    "eslint.validate": [
-        "javascript",
-        "typescript"
-    ],
-    "editor.tabSize": 2
-    
-}
+    "editor.detectIndentation": false,
+    "editor.insertSpaces": true,
+    "editor.tabSize": 2,
+    "files.insertFinalNewline": true,
+    "typescript.preferences.quoteStyle": "single",
+    "eslint.format.enable": true,
+    "editor.defaultFormatter": "vscode.typescript-language-features"
+  }
+
