+++ src/lib/utils.ts
@@ -1,4 +1,23 @@
 
+
+export const constructHref = (
+  baseUrl: string, 
+  params?: Record<string, string>
+) => {
+  const url = new URL(baseUrl);
+    
+  if (!params) return url.href;
+
+  Object
+    .entries(params)
+    .forEach(([key, value]) => {
+      if (typeof value !== 'string') throw `Invalid parameter type ${typeof value}`;
+      url.searchParams.set(key, value);
+    });
+
+  return url.href;
+};
+
 export const unix = {
   fromDate: (date: Date) => {
     return `${Math.floor(date.valueOf() / 1000)}`
@@ -12,12 +31,12 @@ export const unix = {
 
 
 
-import { zonedTimeToUtc, utcToZonedTime } from 'date-fns-tz'
+// import { zonedTimeToUtc, utcToZonedTime } from 'date-fns-tz'
 
-const stripTimezone = () => {
-  return zonedTimeToUtc(date, timeZone)
-}
+// const stripTimezone = () => {
+//   return zonedTimeToUtc(date, timeZone)
+// }
 
-const setTimezone = () => {
-  return utcToZonedTime(date, timeZone)
-}
+// const setTimezone = () => {
+//   return utcToZonedTime(date, timeZone)
+// }
